Various updates for C++20 language features and misc fixes
* Imagine: Use C++20 concepts to improve type constraints on many template functions, core concepts are defined in util/concepts.hh
* Imagine: Replace TextMenuItem::makeSelectDelegate() with MenuItemSelectDelegate type using concepts to add additional function signatures and eliminate various wrapper functions
* Imagine: Integrate fmtlib to replace printf-style formatting for non-logger functions, additional helper functions are in util/format.hh
* Imagine: Add utf16String as a std::u16string wrapper with constructors for converting UTF-8 to UTF-16
* Imagine: Combine definitions of util/builtins.h into util/utility.h
* Imagine: Update docs with new compiler requirements
* Imagine: Remove the unneeded override include directory
* Imagine: Replace usage of NonCopyable base class with simpler deletion of move assignment operator
* Imagine: Make all virtual destructors default
* EmuFramework: Use an empty FileIO instead of GenericIO in !handlesGenericIO case of EmuSystem::loadGameFromPath() so it can be safely accessed

+++ EmuFramework/src/BundledGamesView.cc
@@ -31,7 +31,7 @@ BundledGamesView::BundledGamesView(ViewAttachParams attach):
 		{
 			return 1;
 		},
-		[this](const TableView &, unsigned idx) -> MenuItem&
+		[this](const TableView &, size_t idx) -> MenuItem&
 		{
 			return game[0];
 		}
